rename generate in render
