+++ api/shared/certificate/render.contract.ts
@@ -13,7 +13,7 @@ export type RepositoryInterface = Required<ParamsInterface>;
 
 export const handlerConfig = {
   service: 'certificate',
-  method: 'generate',
+  method: 'render',
 };
 
 export const signature = `${handlerConfig.service}:${handlerConfig.method}`;
